+++ src/plugins/irc/irc-protocol.c
@@ -2644,16 +2644,14 @@ IRC_PROTOCOL_CALLBACK(notice)
             notify_private = 0;
             if (ctxt->server->is_connected
                 && ctxt->nick
-                && (weechat_strcasecmp (ctxt->nick, "nickserv") != 0)
-                && (weechat_strcasecmp (ctxt->nick, "chanserv") != 0)
-                && (weechat_strcasecmp (ctxt->nick, "memoserv") != 0))
+                && irc_config_notice_nick_notify (ctxt->nick))
             {
                 /*
                  * add tag "notify_private" only if:
                  *   - server is connected (message 001 already received)
                  * and:
-                 *   - notice is from a non-empty nick different from
-                 *     nickserv/chanserv/memoserv
+                 *   - notice is from a non-empty nick not present in option
+                 *     irc.look.notice_nicks_disable_notify
                  */
                 notify_private = 1;
             }